refactor: decompose ui_components.py and data_models.py into focused modules
Decompose two monolithic files as part of Issue #1303:

ui_components.py (1333 -> 65 lines):
- startup.py (267 lines): StartupResults, GolfSplashScreen, AsyncStartupWorker
- model_card.py (291 lines): DraggableModelCard, MODEL_IMAGES
- docker_dialog.py (155 lines): DockerCheckThread, EnvironmentDialog
- settings_dialog.py (527 lines): SettingsDialog
- help_dialogs.py (175 lines): HelpDialog, LayoutManagerDialog, ContextHelpDock

data_models.py (1363 -> 56 lines):
- geometry.py (432 lines): Vector3, Quaternion
- skeleton.py (282 lines): JointState, ForceVector
- golf_state.py (383 lines): ClubState, SwingMetrics, BallState, etc.
- data_frame.py (290 lines): UnrealDataFrame

Both original files preserved as re-export facades for backward compat.
All 220 unreal_integration tests pass. All files under 800 lines.
